fix: Measurement methods have been moved to the TaskStudent class.
