+++ lab/lab2/student/TaskStudent.java
@@ -26,30 +26,95 @@ package student;/*Создать класс Student с полями:
 import java.util.*;
 
 public class TaskStudent {
+
+    protected static void measureListPerformance(List<DataTypeStudent> list, DataTypeStudent newDataTypeStudent) {
+        int middle = list.size() / 2;
+        long start, end;
+
+        start = System.nanoTime();
+        list.add(newDataTypeStudent);
+        end = System.nanoTime();
+        System.out.println(" Добавление в конец: " + (end - start));
+        start = System.nanoTime();
+        list.add(0, newDataTypeStudent);
+        end = System.nanoTime();
+        System.out.println("Добавление в начало: " + (end - start));
+        start = System.nanoTime();
+        list.remove(list.size() - 1);
+        end = System.nanoTime();
+        System.out.println("Удаление последнего: " + (end - start));
+        start = System.nanoTime();
+        list.remove(0);
+        end = System.nanoTime();
+        System.out.println("Удаление первого: " + (end - start));
+        start = System.nanoTime();
+        list.get(middle);
+        end = System.nanoTime();
+        System.out.println("Get центрального: " + (end - start));
+        start = System.nanoTime();
+        list.get(list.size() - 1);
+        end = System.nanoTime();
+        System.out.println("Get последнего: " + (end - start));
+    }
+
+    protected static void measureSetPerformance(Set<DataTypeStudent> set, DataTypeStudent newDataTypeStudent) {
+        long start, end;
+
+        start = System.nanoTime();
+        set.add(newDataTypeStudent);
+        end = System.nanoTime();
+        System.out.println("Добавление (HashSet): " + (end - start));
+        start = System.nanoTime();
+        set.remove(newDataTypeStudent);
+        end = System.nanoTime();
+        System.out.println("Удаление (HashSet): " + (end - start));
+    }
+
+    protected static void measureMapPerformance(Map<Long, DataTypeStudent> map, DataTypeStudent newDataTypeStudent) {
+        long start, end;
+        int middle = map.size() / 2;
+        start = System.nanoTime();
+        map.put(newDataTypeStudent.getId(), newDataTypeStudent);
+        end = System.nanoTime();
+        System.out.println("Добавление (HashMap): " + (end - start));
+        start = System.nanoTime();
+        map.remove(newDataTypeStudent.getId());
+        end = System.nanoTime();
+        System.out.println("Удаление (HashMap): " + (end - start));
+        start = System.nanoTime();
+        map.get(middle);
+        end = System.nanoTime();
+        System.out.println("Get центрального (HashMap): " + (end - start));
+        start = System.nanoTime();
+        map.get(map.size());
+        end = System.nanoTime();
+        System.out.println("Get последнего (HashMap): " + (end - start));
+    }
+
     public static void main(String[] args) {
-        int size =10_000_000;
-        ArrayList<DataTypeStudent> arrayList = new ArrayList<>();
-        List<DataTypeStudent> linkedList = new LinkedList<>();
-        Set<DataTypeStudent> hashSet = new HashSet<>();
-        Map<Long, DataTypeStudent> hashMap = new HashMap<>();
-        for (long i = 0; i < size; i++) {
-            DataTypeStudent s = new DataTypeStudent(i, "Student" + i);
-            arrayList.add(s);
-            linkedList.add(s);
-            hashSet.add(s);
-            hashMap.put(i, s);
-        }
-        DataTypeStudent newDataTypeStudent = new DataTypeStudent(10_000_001L, "NewStudent");
-        System.out.println("Количество элементов: " + size);
-        System.out.println("Измерение времени (мс):\n");
-        System.out.println("==ArrayList==");
-        DataTypeStudent.measureListPerformance(arrayList, newDataTypeStudent);
-        System.out.println("==LinkedList++");
-        DataTypeStudent.measureListPerformance(linkedList, newDataTypeStudent);
-        System.out.println("==HashSet");
-        DataTypeStudent.measureSetPerformance(hashSet, newDataTypeStudent);
-        System.out.println("==HashMap==");
-        DataTypeStudent.measureMapPerformance(hashMap, newDataTypeStudent);
+            int size =10_000_000;
+            ArrayList<DataTypeStudent> arrayList = new ArrayList<>();
+            List<DataTypeStudent> linkedList = new LinkedList<>();
+            Set<DataTypeStudent> hashSet = new HashSet<>();
+            Map<Long, DataTypeStudent> hashMap = new HashMap<>();
+            for (long i = 0; i < size; i++) {
+                DataTypeStudent s = new DataTypeStudent(i, "Student" + i);
+                arrayList.add(s);
+                linkedList.add(s);
+                hashSet.add(s);
+                hashMap.put(i, s);
+            }
+            DataTypeStudent newDataTypeStudent = new DataTypeStudent(10_000_001L, "NewStudent");
+            System.out.println("Количество элементов: " + size);
+            System.out.println("Измерение времени (мс):\n");
+            System.out.println("==ArrayList==");
+            measureListPerformance(arrayList, newDataTypeStudent);
+            System.out.println("==LinkedList++");
+            measureListPerformance(linkedList, newDataTypeStudent);
+            System.out.println("==HashSet");
+            measureSetPerformance(hashSet, newDataTypeStudent);
+            System.out.println("==HashMap==");
+            measureMapPerformance(hashMap, newDataTypeStudent);
 
     }
 }
